gameクラスの作成
